*** empty log message ***
git-svn-id: https://svn.code.sf.net/p/lass/code/lass/trunk@337 e9412239-ebb7-49aa-8f4a-f9aa784616d7

+++ src/test/test_num_matrix_vector.inl
@@ -234,8 +234,8 @@ void testNumMatrix()
 	LASS_TEST_CHECK(!b.isIdentity());
 	LASS_TEST_CHECK(!b.isSquare());
 	LASS_TEST_CHECK(!b.isDiagonal());
-	LASS_TEST_CHECK_EQUAL(b.rows(), 3);
-	LASS_TEST_CHECK_EQUAL(b.columns(), 4);
+	LASS_TEST_CHECK_EQUAL(b.rows(), static_cast<size_t>(3));
+	LASS_TEST_CHECK_EQUAL(b.columns(), static_cast<size_t>(4));
 	LASS_TEST_CHECK_EQUAL(b, b);
 }
